Add deletedReviewRating in DeleteReviewResDto, Add reviewCount in Place, Refactor modifying rating and review count logic adding setRatingAndReviewCount()

+++ app/src/main/java/com/sju18001/petmanagement/ui/map/review/ReviewViewModel.kt
@@ -6,5 +6,5 @@ import androidx.lifecycle.ViewModel
 class ReviewViewModel : ViewModel() {
     var placeId: ObservableField<Long> = ObservableField(-1)
     var rating: ObservableField<Double> = ObservableField(0.0)
-    var reviewCount: ObservableField<Int> = ObservableField(0)
+    var reviewCount: ObservableField<Long> = ObservableField(0)
 }
